Enable JRuby support
Requiring the mysql2 gem makes it impossible
to run these useful patches without running mri.

We would like to introduce this library to a JRuby
project, and so the goal is to make it possible to
skip including mysql2 as a dependency in cases where
we can detect a Java runtime.